fix: bigint 列へのパラメータを ::int で 32bit に絞らず ::bigint に揃える
rich_documents / course_chapters の revision・schema_version・sort_order は
いずれも bigint（information_schema で実測）。SQL 側が ::int（int4）で
キャストしていたため sqlc も int32 パラメータを生成し、int32 を超える値は
エラーも出ないまま負数へ巻き戻って保存されていた（実測: 2147483648 → -2147483648）。

::bigint に直して再生成し、呼び出し側の int32() 変換も int64() に揃える。

courses.sort_order の ::int は schema.sql の宣言（integer）と一致しているため
ここでは触らない（宣言と実型の食い違いは別途）。

+++ backend/internal/adapter/persistence/queries/course_chapter.sql
@@ -43,9 +43,9 @@ VALUES (
   sqlc.arg(course_id),
   sqlc.arg(created_by_user_id),
   sqlc.arg(title),
-  COALESCE(NULLIF(sqlc.arg(revision)::int, 0), 1),
-  COALESCE(NULLIF(sqlc.arg(schema_version)::int, 0), 1),
-  COALESCE(NULLIF(sqlc.arg(sort_order)::int, 0), 100),
+  COALESCE(NULLIF(sqlc.arg(revision)::bigint, 0), 1),
+  COALESCE(NULLIF(sqlc.arg(schema_version)::bigint, 0), 1),
+  COALESCE(NULLIF(sqlc.arg(sort_order)::bigint, 0), 100),
   sqlc.arg(is_published),
   sqlc.arg(created_at),
   sqlc.arg(updated_at)
